refactor: refactor notes logic and fix styling
