+++ frontend/src/components/Textfield.tsx
@@ -22,7 +22,6 @@ type BaseProps<T extends FieldValues> = {
   defaultValue?: string;
   className?: string;
   icon?: React.ReactNode;
-  iconClassName?: string;
   mode?: "filled" | "outlined";
   registerOptions?: RegisterOptions;
 };
@@ -49,7 +48,6 @@ export function Textfield<T extends FieldValues>({
   handleInputChange,
   className,
   icon,
-  iconClassName,
   type = "text",
   defaultValue = "",
   mode = "outlined",
@@ -76,9 +74,7 @@ export function Textfield<T extends FieldValues>({
           className,
         )}
       >
-        {icon ? (
-          <span className={cn("grid place-items-center pl-1", iconClassName)}>{icon}</span>
-        ) : null}
+        {icon ? <span className={"grid place-items-center pl-1"}>{icon}</span> : null}
         <input
           {...register(name as Path<T>, registerOptions)}
           className="focus-visible:out w-full appearance-none bg-inherit px-2 placeholder-pia_accent outline-none"
