+++ frontend/src/pages/notes.tsx
@@ -1,10 +1,11 @@
-import { useContext, useEffect, useState } from "react";
+import { useContext } from "react";
 
-import { getAllProgressNotes } from "@/api/progressNotes";
 import { Student } from "@/api/students";
 import LoadingSpinner from "@/components/LoadingSpinner";
 import NotePreview from "@/components/ProgressNotes/NotePreview";
 import NotesSelectionList from "@/components/ProgressNotes/NotesSelectionList";
+import { useProgressNotes } from "@/components/ProgressNotes/hooks/useProgressNotes";
+import { useStudentNotes } from "@/components/ProgressNotes/hooks/useStudentNotes";
 import { ProgressNote } from "@/components/ProgressNotes/types";
 import { UserContext } from "@/contexts/user";
 import { useRedirectToLoginIfNotSignedIn } from "@/hooks/redirect";
@@ -24,169 +25,42 @@ export type ViewMode = "list" | "view" | "edit" | "add";
 function Notes() {
   useRedirectToLoginIfNotSignedIn();
   const { firebaseUser } = useContext(UserContext);
-
-  const [firebaseToken, setFirebaseToken] = useState<string>("");
-  const [filteredStudents, setFilteredStudents] = useState<StudentWithNotes[]>([]);
-  const [allStudents, setAllStudents] = useState<StudentWithNotes[]>([]);
-  const [selectedStudent, setSelectedStudent] = useState<StudentWithNotes>({} as StudentWithNotes);
-  const [selectedNote, setSelectedNote] = useState<ProgressNote>({} as ProgressNote);
-  const [mobileView, setMobileView] = useState<"studentList" | "studentDetails">("studentList");
   const { isMobile } = useWindowSize();
 
-  // Used for undoing note deletion
-  const [deletedNote, setDeletedNote] = useState<ProgressNote | undefined>(undefined);
-  // List mode shows all notes for the selected student
-  // View, edit, add modes show a single note
-  const [noteMode, setNoteMode] = useState<ViewMode>("list");
-  const [allProgressNotes, setAllProgressNotes] = useState<
-    Record<string, ProgressNote> | undefined
-  >(undefined);
+  // Manages fetching and state of progress notes
+  const { firebaseToken, allProgressNotes, setAllProgressNotes } = useProgressNotes(firebaseUser);
 
-  useEffect(() => {
-    if (firebaseUser) {
-      firebaseUser
-        ?.getIdToken()
-        .then((token) => {
-          setFirebaseToken(token);
-          getAllProgressNotes(token).then(
-            (result) => {
-              if (result.success) {
-                const sortedData = result.data.sort(
-                  (a, b) =>
-                    new Date(b.dateLastUpdated).getTime() - new Date(a.dateLastUpdated).getTime(),
-                );
-                const progressNotes = sortedData.reduce<Record<string, ProgressNote>>(
-                  (obj, note) => {
-                    obj[note._id] = note;
-                    return obj;
-                  },
-                  {},
-                );
-                setAllProgressNotes(progressNotes);
-              }
-            },
-            (error) => {
-              console.log(error);
-            },
-          );
-        })
-        .catch((error) => {
-          console.error(error);
-        });
-    }
-  }, [firebaseUser]);
+  // Manages state of student selection and note interactions
+  const {
+    filteredStudents,
+    setFilteredStudents,
+    allStudents,
+    selectedStudent,
+    mobileView,
+    noteMode,
+    setNoteMode,
+    selectedNote,
+    setSelectedNote,
+    handlers,
+  } = useStudentNotes(allProgressNotes, setAllProgressNotes);
 
-  const updateStudentProgressNotes = ({ action, noteData }: HandleNoteUpdate) => {
-    setFilteredStudents((prevStudents) => {
-      const updatedStudents = prevStudents.map((student) => {
-        if (student._id === noteData.studentId) {
-          let updatedProgressNotes;
-
-          if (action === "delete") {
-            updatedProgressNotes = student.progressNotes.filter(
-              (prevNote) => prevNote._id !== noteData._id,
-            );
-          } else if (action === "edit") {
-            updatedProgressNotes = student.progressNotes.map((prevNote) =>
-              prevNote._id === noteData._id ? noteData : prevNote,
-            );
-          } else if (action === "add") {
-            updatedProgressNotes = [...student.progressNotes, noteData];
-            updatedProgressNotes.sort(
-              (a, b) =>
-                new Date(b.dateLastUpdated).getTime() - new Date(a.dateLastUpdated).getTime(),
-            );
-          }
-
-          return {
-            ...student,
-            progressNotes: updatedProgressNotes ?? student.progressNotes,
-          };
-        }
-        return student;
-      });
-
-      setAllStudents(updatedStudents);
-      return updatedStudents;
-    });
-  };
-
-  const handleSelectStudent = (studentData: StudentWithNotes) => {
-    if (studentData._id !== selectedNote.studentId) {
-      if (isMobile) {
-        setMobileView("studentDetails");
-      }
-      setNoteMode("list");
-      setSelectedNote({} as ProgressNote);
-    }
-    setSelectedStudent(studentData);
-  };
-
-  const handleNoteUpdate = ({ action, noteData }: HandleNoteUpdate) => {
-    updateStudentProgressNotes({ action, noteData });
-    if (action === "delete" && allProgressNotes) {
-      const updatedNotes = Object.keys(allProgressNotes).reduce<Record<string, ProgressNote>>(
-        (newNotes, id) => {
-          if (id !== noteData._id) {
-            newNotes[id] = allProgressNotes[id];
-          }
-          return newNotes;
-        },
-        {},
-      );
-
-      setAllProgressNotes(updatedNotes);
-    } else if (action === "edit" || action === "add") {
-      setAllProgressNotes((prevNotes) => {
-        const updatedNotes = { ...prevNotes };
-        updatedNotes[noteData._id] = noteData;
-        return updatedNotes;
-      });
-    }
-  };
-
-  // Used to update state after note is deleted. This cannot run immediately because the dialog will otherwise close before the user can undo the deletion
-  const handleFinishDelete = () => {
-    setNoteMode("list");
-    setSelectedNote({} as ProgressNote);
-  };
-
-  const handlePopulatingStudents = (students: StudentWithNotes[]) => {
-    setSelectedStudent(students[0]);
-    setAllStudents(students);
-    setFilteredStudents(students);
-  };
-
-  const handleMobileBack = () => {
-    setMobileView("studentList");
-  };
-
-  const handleFilterQuery = (query: string) => {
-    if (query === "") {
-      setFilteredStudents(allStudents);
-    } else {
-      setFilteredStudents(
-        allStudents.filter((student) => {
-          return (
-            `${student.student.firstName} ${student.student.lastName}`
-              .toLowerCase()
-              .includes(query.toLowerCase()) ||
-            student.progressNotes.some((note) =>
-              note.content.toLowerCase().includes(query.toLowerCase()),
-            )
-          );
-        }),
-      );
-    }
-  };
+  const {
+    handleSelectStudent,
+    handlePopulatingStudents,
+    handleFilterQuery,
+    handleNoteUpdate,
+    handleMobileBack,
+    handleFinishDelete,
+  } = handlers;
 
   return (
-    <section className="flex h-full flex-col text-sm ">
-      <h1 className={"mb-5 font-[alternate-gothic] text-4xl"}>Progress Notes</h1>
+    <section className="flex h-full flex-col text-[12px] sm:text-sm">
+      <h1 className={"mb-5 font-[alternate-gothic] text-2xl lg:text-4xl "}>Progress Notes</h1>
       {!firebaseUser || allProgressNotes === undefined ? (
         <LoadingSpinner />
       ) : (
         <div className="flex h-full gap-5 overflow-hidden">
+          {/* Show both components on Desktop */}
           {!isMobile ? (
             <>
               <NotesSelectionList
@@ -203,13 +77,13 @@ function Notes() {
                 allProgressNotes={allProgressNotes}
                 selectedStudent={selectedStudent}
                 firebaseToken={firebaseToken}
-                handlers={{ handleFilterQuery, handleNoteUpdate }}
+                handlers={{ handleFilterQuery, handleNoteUpdate, handleFinishDelete }}
                 noteProps={{ noteMode, setNoteMode, selectedNote, setSelectedNote }}
-                deleteProps={{ deletedNote, setDeletedNote, handleFinishDelete }}
               />
             </>
           ) : (
             <>
+              {/* Show only one component at a time on Mobile */}
               {mobileView === "studentList" ? (
                 <NotesSelectionList
                   allProgressNotes={allProgressNotes}
@@ -226,9 +100,13 @@ function Notes() {
                   allProgressNotes={allProgressNotes}
                   selectedStudent={selectedStudent}
                   firebaseToken={firebaseToken}
-                  handlers={{ handleFilterQuery, handleNoteUpdate, handleMobileBack }}
+                  handlers={{
+                    handleFilterQuery,
+                    handleNoteUpdate,
+                    handleMobileBack,
+                    handleFinishDelete,
+                  }}
                   noteProps={{ noteMode, setNoteMode, selectedNote, setSelectedNote }}
-                  deleteProps={{ deletedNote, setDeletedNote, handleFinishDelete }}
                   isMobile={isMobile}
                 />
               )}
